feat: move the parsing of action String to IAction from the Factory to the manager
