+++ src/main/kotlin/com/github/recraftedcivilizations/tasks/TaskManager.kt
@@ -1,7 +1,10 @@
 package com.github.recraftedcivilizations.tasks
 
 import com.github.recraftedcivilizations.dPlayer.DPlayerManager
+import com.github.recraftedcivilizations.tasks.actions.Actions
+import com.github.recraftedcivilizations.tasks.actions.IAction
 import net.milkbowl.vault.economy.Economy
+import java.lang.IllegalArgumentException
 
 /**
  * @author DarkVanityOfLight
@@ -22,8 +25,18 @@ class TaskManager(val econ: Economy, val dPlayerManager: DPlayerManager) {
      * @param action A list of actions this task consist of
      * @param description A description of this task
      */
-    fun createTask(name: String, income: Int, xp: Int, action: List<String>, description: String){
-        val task = TaskFactory.createTask(name, income, xp, action, description, dPlayerManager, econ)
+    fun createTask(name: String, income: Int, xp: Int, actions: List<String>, description: String){
+        val parsedActions = emptyList<IAction>().toMutableList()
+
+        for (action in actions){
+            try{
+                val parsedAction = Actions.valueOf(action)
+                parsedActions.plus(parsedAction)
+
+            } catch (e: IllegalArgumentException){}
+
+        }
+        val task = TaskFactory.createTask(name, income, xp, parsedActions, description, dPlayerManager, econ)
         tasks.add(task)
     }
